requirements.txt
# Munin engine Python dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0
pytest>=7.0.0
hypothesis>=6.0.0

# Intelligence layers (Layers 2-7)
torch>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
